Move cookie checking functions to auth.js
